fix(directory): 파일명 개선
